requirements.txt
openai-whisper
tkinterdnd2
